% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{aggregate_counts}
\alias{aggregate_counts}
\title{Aggregate Counts}
\usage{
aggregate_counts(data)
}
\arguments{
\item{data}{A data frame of counts, typically from \code{get_counts()}.}
}
\value{
A data frame with aggregated counts.
}
\description{
Aggregates count data. This sums counts across directions (and any other non-grouping columns)
for the same ID, time, and class.
}
\details{
It attempts to intelligently aggregate by sensor by checking for a \code{sensor_name} column (created by \code{get_counts()}).
If \code{sensor_name} is missing but \code{name} is present, it will derive \code{sensor_name} from \code{name}.
If neither is present, it aggregates by the existing \code{id}.
}
