% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{aggregate_metadata}
\alias{aggregate_metadata}
\title{Aggregate Metadata}
\usage{
aggregate_metadata(metadata, centroids = TRUE)
}
\arguments{
\item{metadata}{An sf object of countline metadata, typically from \code{get_countline_metadata()}.}

\item{centroids}{Logical. If TRUE (default), converts the aggregated geometry to centroids.}
}
\value{
An sf object with aggregated metadata.
}
\description{
Aggregates countline metadata by simplifying the countline names (extracting the sensor ID)
and returning one row per sensor.
It combines original IDs and names into comma-separated strings and counts the number of countlines per sensor.
}
