% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{get_countline_counts}
\alias{get_countline_counts}
\title{Get Countline Counts}
\usage{
get_countline_counts(
  countline_ids,
  from,
  to,
  classes = NULL,
  time_bucket = "24h",
  wait = 1
)
}
\arguments{
\item{countline_ids}{Vector of countline IDs.}

\item{from}{Start timestamp.}

\item{to}{End timestamp.}

\item{classes}{Vector of classes to include.}

\item{time_bucket}{Time bucket size (e.g. "1h", "5m"). Defaults to "24h".}

\item{wait}{Seconds to wait between API requests. Defaults to 1.}
}
\value{
Data frame of counts.
}
\description{
Get Countline Counts
}
