% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed.R
\name{get_countline_speed}
\alias{get_countline_speed}
\title{Get Countline Speeds}
\usage{
get_countline_speed(countline_ids, from, to, time_bucket = "24h")
}
\arguments{
\item{countline_ids}{Vector of countline IDs.}

\item{from}{Start timestamp.}

\item{to}{End timestamp.}

\item{time_bucket}{Time bucket size (e.g. "1h", "5m"). Defaults to "24h".}
}
\value{
Data frame of speeds.
}
\description{
Get Countline Speeds
}
