% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journey_times.R
\name{get_journey_times}
\alias{get_journey_times}
\title{Get Journey Times}
\usage{
get_journey_times(origin_id, dest_id, from, to)
}
\arguments{
\item{origin_id}{Origin countline ID.}

\item{dest_id}{Destination countline ID.}

\item{from}{Start timestamp.}

\item{to}{End timestamp.}
}
\value{
Data frame of journey times.
}
\description{
Get Journey Times
}
