% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_simplify}
\alias{name_simplify}
\title{Simplify Countline Name}
\usage{
name_simplify(name)
}
\arguments{
\item{name}{A character vector of countline names.}
}
\value{
A character vector of simplified names.
}
\description{
Extracts the unique sensor component from a countline name.
Assumes the format "SensorID_Location_...".
}
\examples{
name_simplify("S38_eastgate_crossing_lpti")
}
